+++ POS/POS/Modules/Payout/Settings/BankSetupConfigSettings.cs
@@ -1,0 +1,9 @@
+﻿using Framework.Infrastructure.Data.Configuration;
+
+namespace POS.Modules.Payout.Settings
+{
+    public class BankSetupConfigSettings : DataConfigItem
+    {
+        public decimal DefaultBankLockupAmount {get; set;}
+    }
+}